+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 # 0.6.0
 
+- Now serializer works with any input type that implements `AsRef<[u8]>`
 - Now deserializer works with any output type that implements `From<Vec<u8>>`
 
 # 0.5.0